move replay-eest scenario
