+++ scenarios/scenarios.go
@@ -17,7 +17,7 @@ import (
 	"github.com/ethpandaops/spamoor/scenarios/factorydeploytx"
 	"github.com/ethpandaops/spamoor/scenarios/gasburnertx"
 	"github.com/ethpandaops/spamoor/scenarios/geastx"
-	"github.com/ethpandaops/spamoor/scenarios/replayeest"
+	replayeest "github.com/ethpandaops/spamoor/scenarios/replay-eest"
 	"github.com/ethpandaops/spamoor/scenarios/setcodetx"
 	erc20bloater "github.com/ethpandaops/spamoor/scenarios/statebloat/erc20_bloater"
 	"github.com/ethpandaops/spamoor/scenarios/storagespam"
